contracts: Fix bug when class overrides __len__.
Previously we were implicitly converting a class instance into a boolean
in _check_function_contracts. This led to an infinite recursion error if
__len__ was overridden (and __bool__ was not), or if __bool__ was
overridden, because those methods would be decorated with contract
checking as well.

Fixed this by changing the implicit boolean conversion into an explicit
`instance is None` check.